+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod logging;
 pub mod models;
 pub mod routes;
 pub mod security;
+pub mod self_update;
 pub mod server_configure;
 pub mod shutdown;
 pub mod stats;